Refactor project structure
This commit will move all source files in a separate src folder.
